chore: remove cyclical dependencies
