+++ apps/ledger-live-mobile/src/screens/SyncOnboarding/TwoStepStepper/useFirstStepCompanionState.ts
@@ -3,8 +3,7 @@ import {
   OnboardingStep as DeviceOnboardingStep,
 } from "@ledgerhq/live-common/hw/extractOnboardingState";
 import { useEffect, useRef } from "react";
-import { FirstStepCompanionStepKey } from "./useCompanionSteps";
-import { SeedPathStatus } from "./FirstStepSyncOnboarding";
+import { FirstStepCompanionStepKey, SeedPathStatus } from "./types";
 import { SeedOriginType } from "@ledgerhq/types-live";
 
 interface UseFirstStepCompanionStateProps {
